Fix double-precision complex typedefs in generated raw layer (#158)
Clang.jl mis-resolves Apple's anonymous `_Complex` typedefs (e.g.
`typedef _Complex double __double_complex_t;`) and emitted the
double-precision complex element types as `ComplexF32` instead of
`ComplexF64` in src/lib/LibAccelerate.jl:

  const __double_complex_t       = ComplexF32  # should be ComplexF64
  const __SPARSE_double_complex  = ComplexF32  # should be ComplexF64

The single-precision typedefs (`__float_complex_t`,
`__SPARSE_float_complex`) correctly map to `ComplexF32` and are left
untouched.

This is currently latent — the idiomatic layer hand-avoids these
wrappers — but it is a corruption landmine: any future migration of a
double-complex symbol (e.g. `vvcosisin` typed `Ptr{__double_complex_t}`,
or `SparseMatrix_Complex_Double.data::Ptr{__SPARSE_double_complex}`)
would silently get half-width buffers.

Two-part, reproducible fix:

1. gen/generate.jl: add a `fix_double_complex_typedefs!` post-process
   step (mirroring the existing `strip_out_of_scope_blas!` BLAS strip)
   that rewrites exactly the two offending `const ... = ComplexF32`
   lines to `ComplexF64`. Targeted and idempotent, so the committed
   output stays deterministic.

2. src/lib/LibAccelerate.jl: apply the same correction by hand to the
   two committed lines so the file matches what the updated generator
   would now produce. No full regen — the runner SDK differs from the
   reference SDK and a full regen would introduce unrelated drift.

Add test/lib_tests.jl (wired into runtests.jl) locking in
`__double_complex_t === ComplexF64`, `__SPARSE_double_complex ===
ComplexF64`, and that the single-precision typedefs stay `ComplexF32`.

Full test suite stays green.

+++ src/lib/LibAccelerate.jl
@@ -2039,7 +2039,7 @@ const DOUBLE_COMPLEX_SPLIT = DSPDoubleSplitComplex
 
 const __float_complex_t = ComplexF32
 
-const __double_complex_t = ComplexF32
+const __double_complex_t = ComplexF64
 
 function vvrecf(arg1, arg2, arg3)
     @ccall libacc.vvrecf(arg1::Ptr{Cfloat}, arg2::Ptr{Cfloat}, arg3::Ptr{Cint})::Cvoid
@@ -2965,7 +2965,7 @@ function Base.propertynames(x::SparseAttributes_t, private::Bool = false)
         end...)
 end
 
-const __SPARSE_double_complex = ComplexF32
+const __SPARSE_double_complex = ComplexF64
 
 struct SparseAttributesComplex_t
     data::NTuple{4, UInt8}
